requirements.txt
# FastAPI and Web Framework
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Database and ORM
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
alembic==1.12.1

# Data Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# External API Integration
httpx==0.25.2
aiohttp==3.9.1

# Schwab API Integration
schwab-package @ git+https://github.com/jkristofgh/schwab-package.git
loguru==0.7.2
pandas>=2.0.0

# Environment and Configuration
python-dotenv==1.0.0
python-multipart==0.0.6

# Testing Framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Notification Libraries
slack-sdk==3.26.1
pygame==2.5.2
numpy==1.26.2

# Performance and Monitoring
structlog==23.2.0